Changing ConnectAsync methods to extensions methods. (#2)
* Changing ConnectAsync methods to extensions methods.

* Changing RxExtensions methods names.

* Backward compatibility

* Move RxExtensions file.

* ... and unify namespace.

* Fix namespace.
