+++ RxSockets/RxSocketClient.cs
@@ -8,12 +8,13 @@ using System.Reactive.Concurrency;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using System.Reactive.Disposables;
+using RxSockets;
 
 #nullable enable
 
 namespace RxSockets
 {
-    public interface IRxSocketClient: IDisposable
+    public interface IRxSocketClient : IDisposable
     {
         bool Connected { get; }
         void Send(byte[] buffer, int offset = 0, int length = 0);
@@ -95,23 +96,14 @@ namespace RxSockets
 
             });
         }
-
+        public static Task<IRxSocketClient> ConnectAsync(IPEndPoint endPoint, int timeout = -1, CancellationToken ct = default) => RxExtensions.ConnectRxSocketClientAsync(endPoint, timeout, ct);
+        public static async Task<IRxSocketClient> ConnectAsync(IPEndPoint endPoint, ILogger<RxSocketClient> logger, int timeout = -1, CancellationToken ct = default) => await RxExtensions.ConnectRxSocketClientAsync(endPoint, logger, timeout, ct);
         public void Send(byte[] buffer, int offset = 0, int length = 0)
         {
             if (length == 0)
                 length = buffer.Length;
             Logger.LogTrace($"Sending {length} bytes.");
             Socket.Send(buffer, offset, length, SocketFlags.None);
-        }
-
-        // static!
-        public static Task<IRxSocketClient> ConnectAsync(IPEndPoint endPoint, int timeout = -1, CancellationToken ct = default)
-            => ConnectAsync(endPoint, NullLogger<RxSocketClient>.Instance, timeout, ct);
-
-        public static async Task<IRxSocketClient> ConnectAsync(IPEndPoint endPoint, ILogger<RxSocketClient> logger, int timeout = -1, CancellationToken ct = default)
-        {
-            var socket = await SocketConnector.ConnectAsync(endPoint, logger, timeout, ct).ConfigureAwait(false);
-            return new RxSocketClient(socket, logger);
         }
 
         public void Dispose()
