+++ RxSockets/RxSocketServer.cs
@@ -8,12 +8,12 @@ using Microsoft.Extensions.Logging.Abstractions;
 using System.Reactive.Disposables;
 using System.Collections.Generic;
 using System.Threading;
-
+using RxSockets;
 #nullable enable
 
 namespace RxSockets
 {
-    public interface IRxSocketServer: IDisposable
+    public interface IRxSocketServer : IDisposable
     {
         IObservable<IRxSocketClient> AcceptObservable { get; }
     }
@@ -27,7 +27,7 @@ namespace RxSockets
         private readonly SocketDisposer Disposer;
         public IObservable<IRxSocketClient> AcceptObservable { get; }
 
-        private RxSocketServer(Socket socket, ILogger logger)
+        internal RxSocketServer(Socket socket, ILogger logger)
         {
             Logger = logger;
             Disposer = new SocketDisposer(socket, logger);
@@ -65,7 +65,7 @@ namespace RxSockets
                     }
                     catch (Exception e)
                     {
-                       Logger.LogTrace("Accept Ended."); // crashes logger
+                        Logger.LogTrace("Accept Ended."); // crashes logger
                         if (!Cts!.IsCancellationRequested && !Disposer.DisposeRequested)
                             Logger.LogInformation(e, "Async Exception.");
                         observer.OnCompleted();
@@ -75,24 +75,8 @@ namespace RxSockets
                 return Disposable.Create(() => Cts?.Cancel());
             });
         }
-
-        public static IRxSocketServer Create(IPEndPoint endPoint, int backLog = 10) =>
-            Create(endPoint, NullLogger<RxSocketServer>.Instance, backLog);
-
-        public static IRxSocketServer Create(IPEndPoint endPoint, ILogger<RxSocketServer> logger, int backLog = 10)
-        {
-            if (endPoint == null)
-                throw new ArgumentNullException(nameof(endPoint));
-            if (backLog < 0)
-                throw new Exception($"Invalid backLog: {backLog}.");
-            logger.LogInformation($"Creating server at EndPoint: {endPoint}.");
-            var socket = Utilities.CreateSocket();
-            socket.Bind(endPoint);
-            socket.Listen(backLog);
-            logger.LogTrace("Listening.");
-            return new RxSocketServer(socket, logger);
-        }
-
+        public static IRxSocketServer Create(IPEndPoint endPoint, int backLog = 10) => RxExtensions.CreateRxSocketServer(endPoint, backLog);
+        public static IRxSocketServer Create(IPEndPoint endPoint, ILogger<RxSocketServer> logger, int backLog = 10) => RxExtensions.CreateRxSocketServer(endPoint, logger, backLog);
         public void Dispose()
         {
             Cts?.Cancel();
